requirements.txt
apeye>=0.0.5
attr_utils>=0.4.0
attrs>=20.2.0
backports-datetime-fromisoformat==1.0.0; python_version < "3.7"
domdf_python_tools>=0.5.1
enum_tools>=0.6.1
typing_extensions>=3.7.4.3
